+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [1.0.2] - 2024-12-25
+
+### Added
+
+- Settings to display/hide inlays
+- Settings to configure dictionary file
+
 ## [1.0.1] - 2024-12-24
 
 ### Added
@@ -15,6 +22,7 @@
 
 - Initial working version
 
-[Unreleased]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/hpalma/intellij-domain-translator/commits/v1.0.0